Crate sp_governed_map

Structs

GovernedMapChangeV1

Type describing a change made to a single key-value pair in the Governed Map.

MainChainScriptsV1

Cardano identifiers necessary for observation of the Governed Map

Enums

GovernedMapInherentDataProvider

Inherent data provider providing the list of Governed Map changes that occured since previous observation.

InherentError

Error type returned when creating or validating the Governed Map inherent

InherentProviderCreationError

Error type returned when creation of GovernedMapInherentDataProvider fails.

Constants

INHERENT_IDENTIFIER

Traits

GovernedMapDataSource

Cardano observability data source API used by GovernedMapInherentDataProvider.

GovernedMapIDPApi

Runtime API exposing data required for the GovernedMapInherentDataProvider to operate.

OnGovernedMappingChange

Handler trait for runtime components which need to react to changes in the Governed Map.